+++ includes/class-vindi-api.php
@@ -26,7 +26,9 @@ class Vindi_API
         'invalid_parameter|card_number'          => 'Número do cartão inválido.',
         'invalid_parameter|registry_code'        => 'CPF ou CNPJ Invalidos',
         'invalid_parameter|payment_company_code' => 'Método de pagamento Inválido',
-        'invalid_parameter|payment_company_id'   => 'Método de pagamento Inválido'
+        'invalid_parameter|payment_company_id'   => 'Método de pagamento Inválido',
+        'invalid_parameter|phones.number'        => 'Número de telefone inválido',
+        'invalid_parameter|phones'               => 'Erro ao cadastrar o telefone'
     );
 
     /**
@@ -341,6 +343,34 @@ class Vindi_API
     }
 
     /**
+     * @param string $code
+     *
+     * @return array|bool|mixed
+     */
+    public function update_customer_phone($id_customer, $phone_number)
+    {
+        $response = $this->request(sprintf('customers/%s', $id_customer),'GET');
+
+        if(empty($response['customer'])) {
+            return false;
+        }
+
+        $phone_data = [
+            'phones' => []
+        ];
+
+        if(empty($response['customer']['phones'])) {
+            array_push($phone_data['phones'], [
+                'phone_type' => 'landline',
+                'number'     => $phone_number,
+            ]);
+            return (boolean) $this->request(sprintf('customers/%s', $id_customer),'PUT', $phone_data);
+        }
+
+        return true;
+    }
+
+    /**
      * @return bool|null
      */
     public function accept_bank_slip()
@@ -500,6 +530,35 @@ class Vindi_API
         }
 
         return $list;
+    }
+
+    /**
+     * @return array
+     */
+    public function get_plan($id)
+    {
+        $response = $this->request('plans/' . $id, 'GET');
+
+        if (empty($response['plan'])) {
+            return false;
+        }
+
+        return $response['plan'];
+    }
+
+    /**
+     * @return int|bool|mixed
+     */
+    public function get_plan_billing_cycles($plan_id)
+    {
+        $plan = $this->get_plan($plan_id);
+
+        if(empty($plan)) {
+            return false;
+        }
+
+        return (int) $plan['billing_cycles'];
+
     }
 
     /**