+++ includes/class-vindi-payment.php
@@ -1,5 +1,4 @@
 <?php
-
 class Vindi_Payment
 {
     /**
@@ -126,6 +125,8 @@ class Vindi_Payment
                 $metadata['carteira_de_identidade'] = $this->order->billing_rg;
         }
 
+        $phone_number = preg_replace('/\D+/', '', '55' . $this->order->billing_phone);
+
         $customer = array(
             'name'          => $name,
             'email'         => $email,
@@ -133,13 +134,24 @@ class Vindi_Payment
             'code'          => $user_code,
             'address'       => $address,
             'notes'         => $notes,
-            'metadata'      => $metadata,
+            'phones'        => [
+                [
+                    'phone_type' => 'landline',
+                    'number'     => $phone_number,
+                ]
+            ],
+            'metadata' => $metadata,
         );
 
         $customer_id = $this->container->api->find_or_create_customer($customer);
 
-        if (false === $customer_id)
+        if(!$this->container->api->update_customer_phone($customer_id, $phone_number)) {
             $this->abort(__('Falha ao registrar o usuário. Verifique os dados e tente novamente.', VINDI_IDENTIFIER ), true);
+        }
+
+        if (false === $customer_id) {
+            $this->abort(__('Falha ao registrar o usuário. Verifique os dados e tente novamente.', VINDI_IDENTIFIER ), true);
+        }
 
         $this->container->logger->log(sprintf('Cliente Vindi: %s', $customer_id));
 
@@ -413,12 +425,20 @@ class Vindi_Payment
     protected function build_product_items_for_subscription($order_item)
     {
         $product_items  = [];
-
         if(empty($order_item)) {
             return $product_items;
         }
 
         $total_discount = $this->order->get_total_discount();
+        $cupons_cycles  = $this->container->cycles_to_discount();
+
+        if(empty($cupons_cycles)) {
+            $discount_cycles = $cupons_cycles;
+        } else {
+            $vindi_plan_id   = $this->get_plan();
+            $plan_cycles     = $this->container->api->get_plan_billing_cycles($vindi_plan_id);
+            $discount_cycles = min($plan_cycles, $cupons_cycles);
+        }
 
         if(!empty($total_discount) && $order_item['type'] == 'product') {
             $order_subtotal      = $this->order->get_subtotal();
@@ -435,7 +455,8 @@ class Vindi_Payment
                 'discounts' => array(
                     array(
                         'discount_type' => 'percentage',
-                        'percentage'    => $discount_percentage
+                        'percentage'    => $discount_percentage,
+                        'cycles'        => $discount_cycles
                     )
                 )
             );